Move some fs-related functions to their own file
Unclutter `util.cc` a bit
